+++ data/src/main/java/team/retum/data/remote/api/student/StudentApi.kt
@@ -1,4 +1,4 @@
-package team.retum.data.remote.api
+package team.retum.data.remote.api.student
 
 import retrofit2.Response
 import retrofit2.http.Body
