+++ data/src/main/java/team/retum/data/remote/datasource/implementation/StudentDataSourceImpl.kt
@@ -1,6 +1,6 @@
 package team.retum.data.remote.datasource.implementation
 
-import team.retum.data.remote.api.StudentApi
+import team.retum.data.remote.api.student.StudentApi
 import team.retum.data.remote.datasource.declaration.StudentsDataSource
 import team.retum.data.remote.request.students.EditProfileImageRequest
 import team.retum.data.remote.request.students.ResetPasswordRequest
