+++ domain/src/main/java/team/retum/domain/enums/Gender.kt
@@ -1,0 +1,8 @@
+package team.retum.domain.enums
+
+enum class Gender(
+    val value: String,
+) {
+    MAN("남"),
+    WOMAN("여"),
+}
